Remove beem references, move get_allowed_accounts, async hive broadcast
